Taking code from previous PR and adjusting routes in order to arrive on introduction carousel and then navigating into different pages with suject in home navbar
