+++ frontend/package.json
@@ -10,16 +10,17 @@
     "fix": "prettier --write . && eslint --fix --ext .jsx ."
   },
   "dependencies": {
-    "@mui/material": "^5.8.3",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
+    "@mui/material": "^5.8.3",
     "axios": "^0.27.2",
     "prop-types": "^15.8.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router-dom": "^6.3.0",
     "styled-components": "^5.3.5",
-    "styled-reset": "^4.4.1"
+    "styled-reset": "^4.4.1",
+    "swiper": "^8.2.4"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.7",
